+++ examples/api/lib/material/floating_action_button/floating_action_button.0.dart
@@ -6,21 +6,38 @@ import 'package:flutter/material.dart';
 
 /// Flutter code sample for [FloatingActionButton].
 
-void main() => runApp(const FloatingActionButtonExampleApp());
+void main() {
+  runApp(const FloatingActionButtonExampleApp());
+}
 
 class FloatingActionButtonExampleApp extends StatelessWidget {
-  const FloatingActionButtonExampleApp({super.key});
+  const FloatingActionButtonExampleApp({ super.key });
 
   @override
   Widget build(BuildContext context) {
-    return const MaterialApp(
-      home: FabExample(),
+    return MaterialApp(
+      theme: ThemeData(useMaterial3: true),
+      home: const FloatingActionButtonExample(),
     );
   }
 }
 
-class FabExample extends StatelessWidget {
-  const FabExample({super.key});
+class FloatingActionButtonExample extends StatefulWidget {
+  const FloatingActionButtonExample({ super.key });
+
+  @override
+  State<FloatingActionButtonExample> createState() => _FloatingActionButtonExampleState();
+}
+
+class _FloatingActionButtonExampleState extends State<FloatingActionButtonExample> {
+  // The FAB's foregroundColor, backgroundColor, and shape
+  static const List<(Color?, Color? background, ShapeBorder?)> customizations = <(Color?, Color?, ShapeBorder?)>[
+    (null, null, null), // The FAB uses its default for null parameters.
+    (null, Colors.green, null),
+    (Colors.white, Colors.green, null),
+    (Colors.white, Colors.green, CircleBorder()),
+  ];
+  int index = 0; // Selects the customization.
 
   @override
   Widget build(BuildContext context) {
@@ -31,9 +48,13 @@ class FabExample extends StatelessWidget {
       body: const Center(child: Text('Press the button below!')),
       floatingActionButton: FloatingActionButton(
         onPressed: () {
-          // Add your onPressed code here!
+          setState(() {
+            index = (index + 1) % customizations.length;
+          });
         },
-        backgroundColor: Colors.green,
+        foregroundColor: customizations[index].$1,
+        backgroundColor: customizations[index].$2,
+        shape: customizations[index].$3,
         child: const Icon(Icons.navigation),
       ),
     );
