--- examples/api/lib/material/floating_action_button/floating_action_button.2.dart
@@ -1,44 +1,0 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-import 'package:flutter/material.dart';
-
-/// Flutter code sample for [FloatingActionButton].
-
-void main() => runApp(const FloatingActionButtonExampleApp());
-
-class FloatingActionButtonExampleApp extends StatelessWidget {
-  const FloatingActionButtonExampleApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      theme: ThemeData(colorSchemeSeed: const Color(0xff6750a4), useMaterial3: true),
-      home: const FabExample(),
-    );
-  }
-}
-
-class FabExample extends StatelessWidget {
-  const FabExample({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('FloatingActionButton Sample'),
-      ),
-      body: const Center(child: Text('Press the button below!')),
-      // An example of the floating action button.
-      //
-      // https://m3.material.io/components/floating-action-button/specs
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          // Add your onPressed code here!
-        },
-        child: const Icon(Icons.add),
-      ),
-    );
-  }
-}
